Refactor: code structures
