+++ .github/workflows/test_coverage.yml
@@ -64,7 +64,8 @@ jobs:
   upload-test-coverage:
     needs: [test]
     if: |
-      github.event_name == 'push'
+      github.repository == 'fluttercandies/flutter_tilt'
+      && github.event_name == 'push'
       && github.ref == 'refs/heads/main'
     runs-on: ubuntu-latest
     steps:
