+++ lib/flutter_tilt.dart
@@ -7,8 +7,11 @@ library flutter_tilt;
 export 'src/config/tilt_config.dart';
 export 'src/config/tilt_light_config.dart';
 export 'src/config/tilt_shadow_config.dart';
-export 'src/data/tilt_data.dart' show TiltDataModel, TiltStreamModel;
 export 'src/enums.dart';
+
+export 'src/models/tilt_data_model.dart' show TiltDataModel;
+export 'src/models/tilt_stream_model.dart' show TiltStreamModel;
+
 export 'src/tilt.dart';
-export 'src/widget/tilt_light.dart';
-export 'src/widget/tilt_shadow.dart';
+export 'src/widgets/tilt_light.dart';
+export 'src/widgets/tilt_shadow.dart';
