+++ lib/src/enums.dart
@@ -111,18 +111,18 @@ enum GesturesType { none, touch, hover, controller, sensors }
 /// Light & Shadow Mode
 /// 光影 Mode
 enum LightShadowMode {
-  /// [No performance risk]
+  /// "No performance risk"
   ///
   /// Use normal shadow effects only for `Tilt.child` without performance degradation.
   ///
   /// ------
   ///
-  /// [无性能风险]
+  /// “无性能风险”
   ///
   /// 仅对 `Tilt.child` 使用普通阴影效果，无性能损耗。
   base,
 
-  /// [Performance risk exists]
+  /// "Performance risk exists"
   ///
   /// Apply a shadow to the entire `Tilt` widget, similar to a projector effect.
   /// will display a shadow that exactly matches the non-transparent parts of the widget.
@@ -134,7 +134,7 @@ enum LightShadowMode {
   ///
   /// ------
   ///
-  /// [有性能风险]
+  /// “有性能风险”
   ///
   /// 对整个 `Tilt` widget 使用类似投影仪效果的阴影，
   /// 会显示与 widget 非透明部分完全一致的阴影。
