+++ lib/src/widgets/tilt_shadow.dart
@@ -5,7 +5,7 @@ import 'package:flutter/widgets.dart';
 import '../config/tilt_light_config.dart';
 import '../config/tilt_shadow_config.dart';
 import '../enums.dart';
-import '../internal/tilt_decoration_mixin.dart';
+import '../internal/mixin/tilt_decoration_mixin.dart';
 import '../utils.dart';
 
 abstract class TiltShadow extends StatelessWidget with TiltDecoration {
