+++ test/internal/tilt_data_test.dart
@@ -1,8 +1,7 @@
 import 'package:flutter/widgets.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:flutter_tilt/src/config/tilt_config.dart';
-import 'package:flutter_tilt/src/data/tilt_data.dart';
-import 'package:flutter_tilt/src/enums.dart';
+import 'package:flutter_tilt/src/internal/tilt_data.dart';
 
 void main() {
   group('TiltData ::', () {
@@ -69,42 +68,6 @@ void main() {
         ),
         reason: 'enableReverse = true',
       );
-    });
-  });
-
-  group('TiltDataModel ::', () {
-    test('hashCode', () {
-      final TiltDataModel tiltDataModel = TiltDataModel(
-        position: const Offset(5, 5),
-        areaProgress: Offset.zero,
-        transform: Matrix4.identity(),
-        angle: Offset.zero,
-      );
-      final TiltDataModel tiltDataModel2 = TiltDataModel(
-        position: const Offset(5, 5),
-        areaProgress: Offset.zero,
-        transform: Matrix4.identity(),
-        angle: Offset.zero,
-      );
-      expect(tiltDataModel, tiltDataModel2);
-      expect(tiltDataModel.hashCode, tiltDataModel2.hashCode);
-    });
-  });
-
-  group('TiltStreamModel ::', () {
-    test('hashCode', () {
-      const TiltStreamModel tiltStreamModel = TiltStreamModel(
-        position: Offset(5, 5),
-        gesturesType: GesturesType.controller,
-        gestureUse: true,
-      );
-      const TiltStreamModel tiltStreamModel2 = TiltStreamModel(
-        position: Offset(5, 5),
-        gesturesType: GesturesType.controller,
-        gestureUse: true,
-      );
-      expect(tiltStreamModel, tiltStreamModel2);
-      expect(tiltStreamModel.hashCode, tiltStreamModel2.hashCode);
     });
   });
 }
