+++ test/tilt_widget/light_shadow_mode_base_widget/tilt_gestures_hover_test.dart
@@ -2,7 +2,7 @@ import 'package:flutter/gestures.dart';
 import 'package:flutter/widgets.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:flutter_tilt/flutter_tilt.dart';
-import 'package:flutter_tilt/src/data/tilt_data.dart';
+import 'package:flutter_tilt/src/internal/tilt_data.dart';
 import 'tilt_widget_base.dart';
 
 void main() {
